feat(engine): extract settings leaf modules to loopover-engine
Continue the settings/ slice extraction (#4879): move the dependency-free pure
leaf global-contributor-cap.ts into packages/loopover-engine, and converge the
two already-twinned pure leaves auto-close-exempt.ts and moderation-rules.ts
onto their engine copies. Each src/ path becomes a thin re-export shim (relative
source path, per src/signals/check-summary.ts), so every consumer and test
resolves unchanged and the single implementation now lives in the engine.

All three are pure (no Env/D1) and at 100% coverage; the engine copies are
byte-identical to the former src/ twins, so behavior is unchanged. The two
converted twins were directory-scanned parity pairs (not named pairs), so the
drift tripwire correctly skips their new shims — no scripts/ change needed.
The remaining settings files (agent-* guarded, and the import-bearing ones)
stay for follow-ups.

Closes #4879
